fix failing test and update types
