+++ app/src/components/sections/CashRemunerationView/__tests__/CRSigne.spec.ts
@@ -74,7 +74,8 @@ vi.mock('@/composables', () => ({
 
 // Mock the constant imports
 vi.mock('@/constant', () => ({
-  USDC_ADDRESS: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48'
+  USDC_ADDRESS: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
+  currentChainId: 137
 }))
 
 // Mock the utils
