+++ app/src/types/index.ts
@@ -20,3 +20,5 @@ export * from './contract-balance'
 export * from './token'
 export * from './trading'
 export * from './safe'
+export * from './safe.mutation'
+export * from './safe.queries'
